Update string-ref to accept <string-cursor>
In SRFI-130 we have a new procedure called string-ref/cursor that could
take either an index or a cursor.

According to the discussion in #567, the preferred approach is to have
one procedure that handles both cursor and index instead of string-ref
handling only index and string-ref/cursor both. In r7rs, it is an error
that string-ref takes something other than an index, but it's up to
implemementation to handle such an error, and we go with a bit liberal
interpretation of that.

With this, string-ref/cursor will be just an alias of string-ref. It's
not added here because string-ref can do the same and having an alias
just pollutes the namespace more. The alias will be added in srfi-130
library later.