Add <string-cursor> type and basic operations (#567) (#569)
These are from SRFI 130. They will be put in srfi-130 library later but
right now (and probably forever) they are part of Gauche builtin
library. This only covers the "Cursor operations" part of the SRFI.

This adds a new cursor type, disjoint from fixnum, to represent the
cursor. We can move forward and backward faster with this, but
converting to index will be slower. Single byte strings have separate
code paths to speed it up because we probably deal with ASCII most of
the time.