Fix <string-pointer>'s size calculation
The field "size" is supposed to cover from "start" to "end", not "index"
to "end". In the multibyte code path we calculate it as "index" to "end"
instead. This problem probably does not show up because most of the time
we execute the single-byte code path.

(Detected by removing the special single-byte code path, moving the
optimization inside forward_pos(), then 'make test' fails)