Improve error messages related to cursors
Printing the cursor itself is not very helpful because even if we print
the pointer inside, the user won't be able to make sense of it. So
instead print the associated string to give some clue.

In some case we might be able to convert the cursor to the byte offset,
relative to the string, which could be helpful. But let's leave that for
later.