srfi-13: update string-index to use cursors
This is to support string-index in srfi-130 which can take either
cursors or indexes. Instead of having separate implementations, we
loosen up srfi-13 a bit to accept both.

This comes with a cost for srfi-13 because <string-cursor> will produce
a lot more intermediate cursor objects, and in the end we will have to
calculate the index while <string-pointer> already has that info at
hand.