Add (most of) srfi-130
srfi-130.scm has been slowly built up over the last few commits. This
makes it a proper module, rexporting procedures that need no
modification.

This is only partial srfi-130 because string-contains-right and
string-replicate are still missing.

Tests are taken from srfi-130 repository. Modified a bit because the scm
file there would not work with an implementation that supports
unicode. New tests (because some times cursors are not used) are added
with test* form instead of the 'or' expression.