srfi-13: update string-trim* to use %string-skip*
The main target here is to replace <string-pointer> with
<string-cursor>. There is some extra cost for going with string-cursor
because in the very last "substring" step, we wouldn need to
count (multibyte) characters in the substring.

Instead of just rewring <string-pointer> code to <string-cursor>, let's
just use %string-skip*. There should be no performance
degradation (besides one or two more procedure calls) and less code to
maintain.