Make <string-cursor> hold offset instead of C pointer
Keeping C pointer in ScmStringCursor was a bad decision [1]. It actually
smells because we can't implement a useful cursor_print(). Because a
cursor is tied to a string anyway, let's just store the byte offset to
beginning of the string.

Most of the code does not change much, just replaced with wrapper macros
to access the offset and convert it to a C pointer like before. String
cursor comparison is greatly simplified because we could just make an
int and compare two ints.

This is important for the next step actually: ScmStringCursor is heavy,
every time we advance a cursor, we actually allocate a new one on
heap. Iterating through a long string with leave us with a litter of
dead cursors waiting to be collected.

So the plan is we have light weight cursors that could fit in an ScmObj,
so no heap allocation. A C pointer won't fit in it, but a small offset
can. With ScmStringCursor also holding an offset, we can handle two
different cursor types the same way.

[1] Moving away from pointers does affect error handling a bit. If it's
a C pointer, you can't accidentally use the pointer made for one string
for another one because range check detects that immediately (unless the
other string is a shared substring). With offset approach, it will go
undetected.