Add "small" string cursor type
ScmStringLargeCursor is expensive to use because it's on heap. After you
have gone through a string of N-chars, we have N cursor objects in heap
mostly waiting to be collected.

To avoid this, let's add a new primitive type that can stay on stack (or
VM registers). Only when the string size is larger than what this type
can hold that we use the large cursor type, which should be very rare.

All cursor procedures are updated to recognize both large and small
types (and prefer to create the small types). The large type is still
printed as "<string-cursor>" while the small type is printed
"#<string-cursor>".

Tested with gc, string-index, gc-stat on a long string. Before this,
:bytes-since-gc grows with the string length. After this,
:bytes-since-gc is more or less around a few kb.

Risk: since the large type is rarely used (and not at all tested in the
test suite), breakages could go unnoticed.