Introduce -fsafe-string-cursors
This is similar to Chibi's -Dsafe-string-cursors. This option forces all
cursors to be large. The large cursor is also slightly larger because it
keeps string body's start pointer in it so that we can check if the
right string is used with it.

This mode is obviously not meant for normal use. Only when you want to
make sure your code does not use invalid cursors.

This can also be used by Gauche devs to make sure large cursor code
works well because usually only small cursors are used (and also the
reason that making large cursors larger does not affect anything).