Update r7rs string-for-each to use cursors (#587)
This is the follow up of #581 which did the same for string-map to avoid
string->list. The "single string" version is definitely better than
before as we produce no temporary objects. The "multiple string"
version is still more complicated...

I learned better this time, so 'any' is used instead of 'filter-map' to
reduce temporary lists. 'string-map' also receives the same treatment.