make gosh -ftest work outside source tree
I was surprised that -ftest has no effect when running outside source
tree. I misused it because the document clearly says "inside". But it's
arguably more convenient to run uninstalled gosh from anywhere.

get_install_dir() is added for real on Linux to get installation
path. It is only used for -ftest, not to make relocatable gauche on
Linux (yet).