Fix regex [[:LOWER:]] [[:UPPER:]] and [[:TITLE:]] under casefold
These express different subsets of letters whose differences are
meaningless under casefold. They are now equivalent of char-set:LC which
covers all letters. This is also consistent with SRE.

With this we have (the last letter is in Lt category)

    ((#/[[:LOWER:]]*/i "σζ\x01C5;") 0) => "σζ\x01C5;"

while before we only match up the first two letters.