line-edit: more robust history loading
Catch and ignore all parsing errors (though we probably should keep an
option to print warnings in case this code breaks?)

To savage as much as possible in the event that this file breaks. Get a
list of strings first and do (read) on each string. If one string
breaks, the others may still be recoverable.