line-edit: do not add duplicated history
If a line is executed repeatedly, just keep one of its lines in
history. This is optional in bash (HISTCONTROL=ignoredups) but we don't
have fine control over line-edit so just stick to useful defaults for
now.

Also ignore empty lines (e.g. by pressing enter repeatedly). We actually
have to press enter twice to get a new prompt, which sounds wrong. But
let's leave that for now.