regexp: support SRE grapheme syntax bog and eog
As discussed in #501 this is a hybrid approach where most of the code is
still in Scheme and it's called up from C regex engine.

When Scm_RegExec() detects a beginning or end of grapheme node [1], it
checks if a grapheme predicate has been made and saved in match_ctx. If
not it calls make-grapheme-predicate to create one.

make-grapheme-predicate is kinda like make-grapheme-cluster-reader. It
gets new grapheme boundary positions from the generator and checks if
the given position is at the boundary or not. All boundary positions are
saved in a closure to avoid going running the generator again and again
on each query.

The position is passed in as a cursor instead of an index because
regexp.c operates on C pointers. Converting C pointers to cursors are
much cheaper than indexes on multibyte strings.

This approach increases match_ctx size (i.e. stack usage) which means we
may hit stack limit sooner. But that's probably not a practical problem.
The predicate also has to be recreated when back tracking. But that's
probably better for safety.

[1] The SRFI does not describe these well. But according to chibi, the
only difference between bog and eog is only bog can match the beginning
of a string, and only eog can match the end. All other grapheme
boundaries in between match both.