Apropos marks 'invisible' symbols
Cf. https://github.com/shirok/Gauche/issues/623
Apropos searches module import/inheritance tree, and it may find
matching symbols that are in them but not directly visible from
the starting module.  It is confusing, for it doesn't give a user
how to make the binding available.

This change marks the binding found but not directly visible with
an asterisk before the module:

```
shiro@scherzo:~/src/Gauche/src$ ./gosh -ftest -r7
gosh[r7rs.user]$ ,a rxmatch
rxmatch                        (*gauche)
rxmatch->string                (*gauche)
rxmatch-after                  (*gauche)
...
```

This still doesn't give a clue what the user should do to use those
bindings (she needs to use gauche.base instead of gauche), but at least
give the clue that it is not visible.

Showing an appropriate module to import will be the next step.