line-edit: consider an empty line "complete" (#629)
%input-complete? is used to determine if an expression is not complete
when \n is typed. If not, the REPL should continue reading instead of
evaluating the expression.

The current condition considers an empty string "not complete"
though. This means if you press Enter right away, e.g. to insert a few
empty lines from the last result, you'll have to press twice for the
REPL to consider "complete" and give you a new prompt.

Change the condition here to consider an empty string "complete" to
avoid that.