Enable and wrap profiler code with MOSTLY_FALSE (#637)
Profiler code is almost always off and better be outside of the core
interpreter loop, higher chance of keeping the entire interpreter in the
instruction cache (and hopefully be nicer to branch predictor too).

MOSTLY_FALSE can be used for this. On gcc-8.2.0 (and Intel CPU), I can
see the original comment is true, run_loop() seems not affected by
current MOSTLY_FALSE "calls".

It does move profiler code (about 8 instructions) out though. It jumps
if profilerRunning is on, to some code at the very bottom of
run_loop. Before MOSTLY_FALSE code, it jumps if profilerRunning is off.

No actual performance measurements are done. I fear this is the kind of
micro optimization that is lost in the noise. Small improvements add up
though, hopefully.