More SRE integration to Gauche
The main purpose of this is to make rxmatch and friends take SRE. We
could make these procedures interpret a list as SRE. But then there's
ambiguity with strings, which could be interpreted as RE strings, or
SREs.

So I though it may be better to just reuse "rx" macro from srfi-115.
This way it's also clearer that it's an SRE:

    (rxmatch (rx (* (/ "az"))) "foo")

And we can void the ambiguity with strings.

This is not as efficient as #// read syntax. But in theory rx macro
could be made to compile the sre at compile time if no unquote is found
in the SRE. I haven't found a way to do it, but if we could then it
benefits srfi-115 as well.