Reorganize files in srfi-159
Keep only public .sld in srfi-159 directory. All other .sld (and related
.scm) are now moved to srfi-159/internal. This makes the structure
clearer (Chibi does not have this problem because it's just (chibi show)
there where everything is visible).

This layout is in line with the implementation in srfi-159 repo (which
was also added by me after I added the code here; and I was still not
happy with the name "inner-base")