editable-reader: let line-edit print the prompt
The procedure show-prompt in line-edit module calls the prompter and
expects some string in the standard output. Instead, editable-reader
"cheats" and prints the prompt by itself when the promptr is called.

Let's do the "right thing" by printing to current-output-port. This is
no-op right now. But later on, line-edit may use a virtual console to
catch commands and replay later. Manipulating the console directly from
outside breaks that.