line-edit: add and use a "replay" virtual console
redisplay procedure in line-edit now writes all console commands to a
virtual console. At the end, the virtual console will replay all the
commands to the real console.

This is of course no-op and pretty useless at this point. The plan is,
we can use this command sequence to see what areas are damaged next time
and only redisplay that part.

The rough idea at this point is, we keep the whole command sequence of
the last display. Then we compare with the new command sequence, then
search back to the last cursor movement and only redraw from then. If
the new sequence is:

    reset-character-attribute
    move-cursor-to y 0
    show-prompt
    clear-to-eos
    ...
    move-cursor-to y x [1]
    ....

Then the new command sequence for partial display is

    move-cursor-to y x
    reset-character-attribute
    clear-to-eos
    (replay all set-character-attribute before [1])
    (replay all commands after [1])

This assumes that we redisplay from top down in the command sequence.

We could even do smarter things to avoid redisplaying everything after
[1] (e.g. usually only one line is updated) but we can stick to
something simple for now.