Preparation of split stack
A split stack is for lightweight continuation capturing.  When a continuation
is captured, we move the stackBase beyond the captured frame, and keep it
from being popped.  We flush the protected frames into heap at the timing
when stack is GC-ed or full continuation is captured.

For simplicity, we maintain invariance that no heap objects point into stack.
So we don't use the lightweight continuation for generic call/cc, but it's
still useful.  The first application we're aiming at is the continuation of
"last error"---by inspecting continuation chain where the error occurs, we
can produce a stack trace, but saving entire chain every time error occurs
would be expensive and wasteful if the chain wouldn't be inpsected.

It's all future plan.  This commit only establishes the cpp symbol
GAUCHE_SPLIT_STACK as a preparation.