Add text.console.framebuffer for flicker-free line-edit
This is a preliminary work to keep all console updates in memory, then
display to a real console in the end. Because the last display is kept,
only damaged positions in the console will be redisplayed this
time. This reduces the flicker caused by clear-to-eos at the beginnning
of each redisplay in line-edit.

The entire code is currently disabled. It's not ready for use
yet. Remaining work includes:

- wide character support
- cursor-down/scroll-up handling (and optimizing too if possible)
- cursor-up/scroll-down handling (if detected somehow)
- clear-to-eos optimization (e.g. M-< C-k or something like that)
- perhaps avoid vectors of conses to reduce GC pressure a bit

Part of #665