console.framebuffer: reduce real cursor movement
Keep track of the last position of the cursor and only issue the escape
sequence to move it if it's not already at where it's supposed to be.
This is more about not sending too much stuff to the real console than
flickering due to cursor movement because the cursor is already hidden.