console.framebuffer: handle cursor-down/scroll-up
It is handled by not doing anything actually. Scrolling up operation
requires knowing all the tiles in the console, but we only know the
tiles that have been updated.

So when we hit this case, we set pass-through flag. Instead of partial
redisplay, draw-framebuffer will call replay all the console methods it
has received, but on the real console.

We should only hit this case when when a new line of the current
expression gets to the end of the screen and the screen starts to scroll
up, which is not that often since most of the time people write
one-liners.

Even if it happens, the next edits do not require scrolling anymore, so
framebuffer can still be activated. The worst case scenario is when an
expression is longer than the screen's height. Any edit in this case
will lead to scrolling up (it's how line-edit does it). Which means
framebuffer is completely disabled.