parser.peg - Add failure procedure in $match1
Sometimes, even the input matches the pattern, you want to do additional
check and let the parser fail if the input doesn't meet the criteria.
Now you can write ($match1 pattern (=> identifier) expr).  IDENTIFIER
is bound to a procedure in EXPR, and if you call the it with a message,
the parser fails with that message.

It is like util.match's (=> FAIL) feature, but in util.match, FAIL
is bound to a continuation that abandones the current clause and
try the next clause.  In $match1, the failure procedure is just
a procedure, so you have to call it at the tail position in EXPR.