Don't check full apply's argument list if possible
The function check_arglist_tail_for_apply() goes to the end of the list
to determine the length, and whether it's circular. In many cases we
don't really need to know its full length, just whether it's above a
certain limit.

Make the function accept an upper limit and terminate early in these
cases. We could be a bit faster in the pathological case where we know
in advance how many number a procedure may take and we have to handle a
list of a zillion elements.

In one case where we have to unpack the entire argument list, there will
be no upper limit.

This could be the first step to fixing issue #686 too.