string->number takes exact? optional argument
If true value is given, it assumes the input is exact unless #i is explicitly
given.

(NB: At this moment, we don't have exact complex, so it the input looks like
complex and exact? is #t, string->number gives up and returns #f.  It will
change once we support exact complex numbers.)