rfc.json - avoid consuming trailing whitespaces
srfi-180 uses ports for the source, and it expects, after one json
item is read, the remainder of input is yet to be read from the port.
If our tokenizer tries to consume trailing wss, it will read one
character ahead after wss, which makes that character dropped
from the port.