line-edit: interpret #\newline as C-m
When line-edit is running normally, the tty is configured so that we
should get C-m instead of #\newline. However if the command prompt is
put in background then back in foreground, the tty settings are lost.

It looks like from now on we get #\newline instead when Enter is
pressed. So if we catch this, consider it the same as C-m. If we manage
to commit the expression, the tty should be reset the next time the
command prompt is ready to accept new input.

If the expression is incomplete, one can only hope the user will end up
typing either Ctrl-C or Ctrl-D. Either way the prompt is finished and
the next prompt resets the tty.

PS. We probably could take the hint that #\newline means the tty is
screwed and reset the tty now (since Enter is one of the most obvious
keys to type when you're not seeing anything). But that might have unseen
consequences.

Related to issue #715