text.line-edit - Let read-line/edit handle SIGCONT
Now read-line/edit installs SIGCONT handler.  When the REPL is stopped
and the shell takes the console back, the console status is reset to the
canonical (line editing) mode.  When REPL resumes the operation, it needs
to detect the situation and set up the console again for the editing.

Since the library don't have total control over each thread's sigmask,
we can't guarantee that the thread calling read-line/edit receives SIGCONT.
So we can't immediately handle it.  However, the signal handler records
the fact, and the next input of the user (the user must type ENTER to
send the input to the process, for the terminal is in canonical mode),
we check the situation and reset the terminal if necessary.  A bit confusing,
but tolerable, I think.

See https://github.com/shirok/Gauche/issues/715