gauche.collection - Add collection methods of bitvectors
Also some fallback methods of uvectors.  Gauche.uvector defines
methods specific to each concrete uvector types, but the user
may only load gauche.collection to iterate over uvectors.
(It can be a source of trouble, though, for the code to be executed
depend on whether gauche.uvector is loaded or not.  We may as well
move coll/seq method definitions on uvector into gauche.collection/sequence).