Fix peek-char and getpos issue
If the port has custom getpos/setpos, the position can be an arbitrary
Scheme object, and we can't adjust offset of ungotten character.
Instead, we cache the position when peek-char is used.  Srfi-181 discusses
this issue in detail.

The current code is quite ugly; we'll refactor and clean up later.

We also need to care virtual i/o port, for we have to reset position
before writing to it if the previous operation is peek.