Tentative commit of adding 'inject' feature to er-macro
Er macro assumes bare symbols left in the macro output is interpreted
in the macro use context.  It doesn't go well with other hygienic macros
where the other macro output (which already has its own context) calls
er-macro.  This experimentally adds eri-macro, which allows a transformer
to explicitly 'inject' identifiers into the macro use environment.

To make this work completely, however, we have to change how bindings
are resolved in the compiler---so the tests are commented out for now.