Drop DISABLE_{GET|SET}POS flag from port
It is introduced so that a custom port can disable positioning without
resetting getpos/setpos callback pointer.  However, it turned out we need
to reset them anyway, for the port tries getpos then seek (or setpos
then seek); if getpos pointer is set to a procedure that does nothing,
seek won't be called.  We need to manage {get|set}pos and seek availability
separately after all.  If that's the case, setting unavailable callback to
NULL is suffice.