Revive <io-invalid-position-error>, dropping mixin approach
We employed <io-invalid-position-error-mixin> to adapt to srfi-192,
for make-i/o-invalid-position-error doesn't take port in the argument.
However, it makes things too complicated.  Now that we have flexibility
in adding more info in <port-error>, we revive <io-invalid-position-error>
that adds position slot.  For srfi-181/192 code, we intercept errors
from set-position! callback anyway, so we can add port info.