Remove unnecessary memory barrier
I believe we don't need one before updating lp->owner; as long as lp->owner
is not zero and lazypair tag is valid, other threads see it as a lazy
pair and someone is working on it.  Once the tag field is overwritten
by ScmObj, other threads see it's an extended pair.