Change :init-function special values of dynamic-load
Previously, #f means 'use the default name', and it's the default value,
but never documented.  Now we want to have an option not to call init
function at all, and :init-function #f makes sense for that purpose.
So we changed :init-function #t to indicate to use the default name,
and makes it default for 'dynamic-load'.  There's little reason the
user want to call dynamic-load directly before, and aut-generated
dynmic-load call never explicitly say :init-function #f, so we presume
this won't break existing code.