Gradually moving actual code in scheme.* module from srfi-*
We suppose people will use scheme.* more instead of srfi-*, so it's
reasonable to have the actual code in scheme.* and make srfi-* extend
them for the backward compatibility.  Also, the output of apropos shows
scheme.* name, which is better than srfi-*.

It's WIP.  This commit moves 5 modules.  More to follow.