peg - Extended $one-of to take list of items
It turned out it's quite useful if we can say ($one-of '("abc" "def"))
or ($one-of '(a b c)) ---the latter is when we're parsing a tokenized
stream.

Unfortunately, $none-of can't be generalized in the same way, for
it is ambigous what to return when ($none-of '("abc" "def")) accepts
input.  This asymmetry is unsettling, so we may consider dropping $none-of
in future.