peg - Revise $or error message generation
When all branches of $or failed, it creates a compound failure and synthesizes
an error message.  Before, if there's a fail-message, we just used it.  What
we had in mind is the use case such as this:

  ($or p1 p2 ... ($fail "comprehensive message"))

In this case, it is reasonable to use the "comprehensive message" part.
However, this use case breaks when any intermediate parsers also return
fail-message.  So we should include failures of p1 p2 ... into the message
by default.

The :else mechanism allows the user code to explicitly discard previous
errors and use the user-provided comprehensive error message.