peg - Enhance $binding
We saw a lot of pattern ($binding ($seq ($: var x) ($: var y) ...) body).
Since the result of entire $seq is ignored, and we know body is a single
form, it is suffice to say ($binding ($: var x) ($: var y) ... body).
Similar syntax can be seen in srfi-42.

Also introduce ($lbinding ...) which is ($lazy ($binding ...)).  This is
frequently appearing pattern, and $binding is already a macro, so why not?