gauche.configure - Fix cf-prog-cxx
This has been broken after we changed cf-arg-var specification.
Now that cf-arg-var sets the variable to "" if there's no environment
variable, and other parts of the program thinks the variable already
has a value.

cf-path-prog and cf-check-prog are also fixed to check if the variable
is a null string.

This is an instance of language model mismatch---for the original configure
language, shell script, it doesn't matter whether the variable has an empty
value or the variable isn't defined.  In Scheme, it does.