Fix inlinable binding handling (wip)
We check the redefinition of inlinable bindings.  However, there's a
legit case that inlinable bindings can be superseded---an inlinable
binding is first registered as a dummy binding during compilation,
in order to expand the inlinable procedure within the same compilation
unit, then, at the runtime, proper definition is inserted.  Hence
we should distinguish the dummy inlinable bindings.