Tweak initialization-time binding transfer
We have some internal modules in the core (gauche.object and gauche.typeutil)
purely for maintainability.  The public API entries defined there are
'transfered' to #<module gauche> at the initialization.  We used to use
a quick hack macrology to do that, but we need more control over binding
flags, so we define a special cproc to do the job.  The cproc is only
available during initializaiton, since such operation should be avoided
in general.