Start adding <stub-type>
Apart from proxy-type and descriptive-type, we need to have types
that bridges Scheme and C; until now, it is only handled when stub file
is generated, and gauche.cgen.type manages it completely separately
from the runtime type system.  However, to utilize stub's type information
at runtime, we need to have the corresponding type instance in the core.