Simplify lcs-edit-list/context API
Drop the "omit if no changes" spec in the hunk format.  It is mainly
for humans to parse context diff output.  Programatically we don't get
anything by omitting it, and the caller needs to handle a special case
which is a disadvantage.

The omission is now handled in diff-report/context, for it produces
output to be read by humans.