cgen - Emit forward-decls right before <procstub> body
When CiSE macros within a <procstub> needs to emit something toplevel,
it pushes them in forward-decls.  We used to emit it in the 'decl' part.  But
that interferes if the code in forward-decls refer to something that
needs to be declared (notably, statically generated data).

So we change to emit forward-delcs in the 'body' part, right before
the main body.  As far as we know, forward-delcs doen't have things
that actually needs to be in the 'decl' part and it should work.
(We may as well change the name to forward-defs, but that'll be another patch).