Adding $CLAMBDA IForm node
To optimize case-lambda according to its calling site, it's better
to handle it explicitly, rather than the current way that delegates
everything to runtime.   This commit is the first step to add $CLAMBDA
node.  This compiles, but $CLAMBDA node isn't used yet.

Note: This changes numberic label of IForm node, so the precompiled
files with inlinable procedures becomes incompatible from the one
precompiled by the previous versions.