Straighten iset transient/persistent management
We go with this for now, but the transient/persistent model may not
be optimal.  For example, freshly created iset can be safely passed
to linear-updating procedure, but the current model forces copying
internal tmap.

The right model might be some kind of reference counting.  Doesn't
need to be precise; need to distinguish whether the internal structure
is ever shared or not.