Improve completion
Add whitespace after the completed word if
the completion returns a single definite result.  It is indicated
by the lister callback returning a string, rather than a list.
Note that we need to distinguish the case when there's only one
candidate but further search is possible (e.g. completing a directory
name when searching the actula file).