Rewrote SCM_HEADER/SCM_INSTANCE_HEADER in CiSE struct definition
We thought we couldn't use them, for simply listing them in
the field list of struct causes them to be rendered as a field
of type ScmObj.  We notice that giving an empty type name ||
suffice our needs.  It's ugly but this technique should rarely be
used so it's ok.  We commented this usage in cise.scm.