Introduce bitvector-{any,every}-value?
... intead of bitvector-{any,every}?.  any? tests if there's
at least one 1, every? tests if all bits are 1.  For the compleness,
we also want notany? (all bits are 0) and notevery? (at least one 0);
we previously introduced bitvector-none? as notany?, but can't think
of a good name for notevery? without using 'not'.
By passing the value to look for, however, we only need any and
every.  It also aligns any/every on other collection types that take
a predicate, so that you can flip the meaaning by complementing the
predicate instead of the entire expression.

To avoid confusion of predicate-taking any?/every?, we name these
any-value?/every-value?.