Change C Continuation protocol
Let the C continuation procedure take the VM pointer so that we can
avoid calling Scm_VM() too often.
When we first designed the protocol, the overhead of calling Scm_VM()
is negligible, so we didn't pass it to the c cont proc.  However, as
the processors gets faster, and the advanced precompilatoin/jit would
generate large number of small c cont procs, it is better to carry
around VM pointer.

The new protocol (we call it PCont, just to avoid conflict with CCont)
also takes the pointer to ScmContFrame.  This is to keep the backward
compatibility; we need an adapter procedure to call the original ccont.

This adapter procedure adds a slight overhead to the original ccont.
We benchmarked with a contrived example that heavily uses VMPushCC,
and observed about 4-5% degradation of speed.  In a hand-written
C code, VMPushCC isn't as frequently used, so we tolerate it for now.

(We also tested to use cont->marker as a flag to distinguish old
protocol and new protocol; the overhead is also around 5% for
the contrived benchmark.  With that code, however, overhead is incurred
on both old and new protocol.  This commit only taxes the old protocol).