Further change of CCont protocol
Eliminate ScmContFrame* argument from the PCont procedure; it is
internal info and shouldn't be exposed to the user code.
In order to pass CCont pointer for the adapter code, we added
a new slot in vm (vm->ccont), which is transiently used within
POP_CONT.  It's a kludge, but it hides ugly details.