Revise %reraise not to refer to ep->floating
The role of %reraise is mererly to tell the VM that the error
handler intends to return to the continuation of raise-continuable
instead of aborting to the outer error handler.  If so, just a simple
flag in VM is suffice---it is set by %reraise, read immediately after
in Scm_VMDefaultExceptionHandler and reset, so no worry about other
error handlers stepping inbetween.

Realted to https://github.com/shirok/Gauche/issues/852