Transition to use prompt tag to mark boundary frame
Instead of using a single boundaryFrameMark to delimit the cont chain,
we'll have multiple continuation prompts.  We use only defaultPromptTag
for the time being, but other prompt tags will be used eventually.

To avoid overhead of checking boundary frame, we use one bit of marker
to indicate boundary frame.  Actually, we expect to get rid of
CONT_SHIFT_MARKER eventually once we revised reset/shift in terms of
continuation prompts.