Update test-vmstack for false positives
Now that user_eval_inner pushes ScmPromptData before the boundary
frame that remains after the frame is popped, the SP values of pre-eval
and post-eval may not match.  We use heuristics to check the
valid outcomes.

It is also noticed that test-vmstack hasn't reflect the test results
in its exit code, so failures have been overlooked for some time.