Fix export of r7rs.import module
We don't need rename, for the macro is defined as 'r7rs-import' within
the module.
I guess this is a remnant during development when we define 'import'
in the module.
It has been harmless, since for R7RS programs 'import' is injected
using (with-module r7rs.import r7rs-import).   However, when you imported
r7rs.import and tried to use r7rs-import, the bug surfaced.