Add call-with-current-expression-name
The expression name is the identifier name the result of tail expression
will be bound.  It is available during pass1 and used to name procedures,
but it lacked flexibility, for

Now the compiler wraps the RHS expression of definition to add the
expression name as a continaution mark.  It can be retrieved at runtime
by call-with-current-expression-name.

Cf. https://github.com/shirok/Gauche/issues/858, but this commit is better
than the suggested one, for using pass1 compiler syntax requires
call-with-current-expression-name visible in the same cmpilation path
of the definition.