Keep dynamic handler entry in CCont of dynamic-wind body
This is a step to change management of dynamci handler chain.

Dynamic handler chain is managed along continuation chain; that is,
when continuation chain is "cut out" and "restored" by delimited
continuation primitives, corresponding dynamic handler chain is also
cut out and restored.  Thus it makes sense that we directly store
a dynamic handler entry in the continuation frame.

This is experimental; let's see if we can get rid of explicit dynamic
handler chain management.