rfc.uuid - Fix parameterization of random source
The previous way to alter %uuid-random-int by the hook of uuid-random-source
dosn't work, for we need to restore %uuid-random-int outside of the
dynamic scope, and also direclty modifying %uuid-random-int isn't thread
safe.

SRFI-27 lacks the API to parameterize random source, so we fall back
to math.mt-random API.

Also added write-object for uuid.