Avoid non-toplevel define-method
Although it's not been explicitly stated, define-method isn't supposed
to be used except toplevel.  It has been expanded into an expression that
adds method closure to gf, so it has worked in non-toplevel context,
but we'll make a change to disallow it.
So far this is the only instance we're aware of non-toplevel define-method
in the Gauche itself.