Let test-module check body of define-method as well
It is done by binding the method body closure to a module-internal
unique toplevel variable, so that test-module can find it.

As a side-effect, define-method is now only allowed at toplevel.  It has
always been the intention, though never explicitly stated.
Existing code that uses define-method in non-toplevel expressions may
get an error.