Change load to return the pathname upon success
This is technically an incompatible change: Previously 'load' returns #t
on success.  We think this is a better interface, for you can get the
path of the actual file (after search paths and suffixes are considered).
As far as the caller treats the return value as a generalized boolean,
it should keep working.