Fix force exception handling
When exception is thrown during evaluating promise body, we need to
abort the continuation upto force and then reraise it from force.
To implemnet it as is, we need a delimited continuation.

Instead, we install a special exception handler during evaluating
promise body, and it just records the condition in the promise
object and reraise it from the outer exception handler of force,
thus we effectively skip the continuation frames upto force.
Next call to force can see the SCM_PROMISE_EXCEPTION state and
raise the condition again.