Revise API of rfc.base64 and rfc.quoted-printable for the consistency
The old API *-{encode|decode}-{string|bytevector} is confusing.
For the encoders, the suffix means type of the argument, while for the
decoders, the suffix indicates the return type.  It is also more convenient
if a single encoder works for both strings and uvector input.

Now the generic encoder is named *-encode-message, and the decoder
*-decode-string-to takes target class, much like digest framework.

The old APIs are kept for the backward compatibility.

This is stemmed from https://github.com/shirok/Gauche/issues/919